refactor: move group role components into own folder
