+++ src/components/Groups/GroupRoles/EditGroupRoleTabs.tsx
@@ -2,10 +2,10 @@ import { Card, Tab, Tabs } from "@mui/material";
 import { useRouter } from "next/router";
 import { SyntheticEvent, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { EditGroupRoleTabsFragment } from "../../apollo/gen";
-import { EditRoleTabNames } from "../../constants/role.constants";
-import { useAboveBreakpoint } from "../../hooks/common.hooks";
-import AddRoleMemberTab from "../Roles/AddRoleMemberTab";
+import { EditGroupRoleTabsFragment } from "../../../apollo/gen";
+import { EditRoleTabNames } from "../../../constants/role.constants";
+import { useAboveBreakpoint } from "../../../hooks/common.hooks";
+import AddRoleMemberTab from "../../Roles/AddRoleMemberTab";
 import DeleteGroupRoleButton from "./DeleteGroupRoleButton";
 import GroupPermissionsForm from "./GroupPermissionsForm";
 import GroupRoleForm from "./GroupRoleForm";
